Can now handle custom tags
