--- _assets/js/netlify/formatDates.js
@@ -1,6 +1,0 @@
-export default function formatDates(date) {
-    const newDate = new Date(date.split(" ")[0].split("-"));
-    const options = { month: "long", day: "numeric", year: "numeric" };
-    const finalDate = new Intl.DateTimeFormat("en-US", options).format(newDate);
-    return `${finalDate}`;
-}
